+++ src/main/java/de/holube/pad/model/DefaultBoard.java
@@ -9,7 +9,7 @@ import java.util.List;
 @EqualsAndHashCode(callSuper = true)
 public class DefaultBoard extends AbstractBoard {
 
-    public static final int[][] BOARD_LAYOUT = new int[][]{
+    public static final byte[][] BOARD_LAYOUT = new byte[][]{
             {0, 0, 0, 0, 0, 0, 1},
             {0, 0, 0, 0, 0, 0, 1},
             {0, 0, 0, 0, 0, 0, 0},
@@ -39,10 +39,10 @@ public class DefaultBoard extends AbstractBoard {
                     {29, 30, 31, -1, -1, -1, -1}
             }
     };
-    protected static final int MAX_KEY;
+    protected static final byte MAX_KEY;
 
     static {
-        MAX_KEY = Arrays.stream(BOARD_MEANING[0])
+        MAX_KEY = (byte) Arrays.stream(BOARD_MEANING[0])
                 .flatMapToInt(Arrays::stream)
                 .summaryStatistics().getMax();
     }
@@ -51,12 +51,12 @@ public class DefaultBoard extends AbstractBoard {
         this(BOARD_LAYOUT, new ArrayList<>(), new ArrayList<>());
     }
 
-    public DefaultBoard(int[][] board, List<Tile> tiles, List<int[][]> tileCumArrays) {
+    public DefaultBoard(byte[][] board, List<Tile> tiles, List<byte[][]> tileCumArrays) {
         super(board, tiles, tileCumArrays, MAX_KEY);
     }
 
     @Override
-    public int[][] getBoardLayout() {
+    public byte[][] getBoardLayout() {
         return BOARD_LAYOUT;
     }
 
@@ -66,12 +66,12 @@ public class DefaultBoard extends AbstractBoard {
     }
 
     @Override
-    protected Board createNewBoard(int[][] newBoard, List<Tile> newTiles, List<int[][]> newTileCumBoards) {
+    protected Board createNewBoard(byte[][] newBoard, List<Tile> newTiles, List<byte[][]> newTileCumBoards) {
         return new DefaultBoard(newBoard, newTiles, newTileCumBoards);
     }
 
     @Override
-    public int getMaxKey() {
+    public byte getMaxKey() {
         return MAX_KEY;
     }
